BIG refactor, small updates and Turnip V2 started
