+++ Cargo.toml
@@ -12,3 +12,6 @@ crate-type = ["cdylib"]
 pest = "2.7.15"
 pest_derive = "2.7.15"
 wasm-bindgen = "0.2.100"
+
+[dev-dependencies]
+pretty_assertions = "1.4.1"
