+++ src/parser/turnip_v2.pest
@@ -1,9 +1,10 @@
-//! Grammar file the the V0 of the Turing Machine language
+//! Grammar file the the V2 of the Turing Machine language
+//! This simply changed `tape_symbol` to use bytes rather than bits.
 //!
 //! This file is used by the pest parser generator to create a struct that
 //! we can then manipulate in the main Rust functions of the library.
 
-// Implicit rules that will apply everywhere - comments
+// Implicit rules that will apply everywhere
 WHITESPACE = _{ "//" ~ (!NEWLINE ~ ANY)* ~ NEWLINE|" "|NEWLINE|"\t" }
 
 // State name
@@ -15,24 +16,30 @@ var_name = @{ LETTER ~ (LETTER | NUMBER)* }
 // Integer
 int = @{ ("+"|"-")? ~ ASCII_DIGIT+ } //can be 0 if we want
 
+// Byte
+byte = @{ ASCII_DIGIT | ASCII_DIGIT ~ ASCII_DIGIT | "1" ~ ASCII_DIGIT ~ ASCII_DIGIT | "2" ~ ( '0'..'4' ~ ASCII_DIGIT | "5" ~ '0'..'5' ) }
+
 // What can be matched by a rule
-tape_symbol = @{ '0'..'1'|"_"|var_name }
+tape_symbol = @{ byte | "_" | var_name }
 
 // Head movement
 head_move = @{ "L"|"R"|"S" }
 
 // Function
 fun_name = @{ UPPERCASE_LETTER~("_"|UPPERCASE_LETTER|ASCII_DIGIT)* }
-function = { fun_name ~ "(" ~ (int|var_name) ~ ")" }
+function = { fun_name ~ "(" ~ (int|var_name)? ~ ")" }
 
 // Function sequence
 fun_sequence = { function | "[]" | "[" ~ function ~ ("," ~ function)* ~ "]" }
 
 // Transition condition
-tape_action = { ("(" ~ tape_symbol ~ "," ~ head_move ~ ")") | fun_sequence }
+tape_action = {
+    "(" ~ tape_symbol ~ "," ~ head_move ~ ")"
+    ~ "," ~ "(" ~ tape_symbol ~ "," ~ head_move ~ ")" | fun_sequence
+}
 
 // Match rule of a state
-state_rule = { "|" ~ tape_symbol ~ "->" ~ tape_action ~ "," ~ state_name }
+state_rule = { "|" ~ tape_symbol ~ "," ~ tape_symbol ~ "->" ~ tape_action ~ "," ~ state_name }
 
 // A state with its match rules
 state = { state_name ~ state_rule+ }
